Harden meal planner circular dependencies
