+++ client/src/components/meal-planner/planner-adaptation/adaptationTypes.ts
@@ -1,23 +1,7 @@
 import type { AdaptiveNutritionIdentity } from '../personality-modeling/personalityTypes';
+import type { LongitudinalPlanningSnapshot } from './longitudinalPlanningTypes';
 
-export type LongitudinalPlanningSnapshot = {
-  id: string;
-  createdAt: string;
-  weekKey: string;
-  completedMeals: number;
-  skippedMeals: number;
-  repeatedMeals: number;
-  prepCompletionRate: number;
-  groceryCompletionRate: number;
-  hydrationAdherence: number;
-  readinessAverage: number;
-  fatigueAverage: number;
-  successfulRelationshipChains: number;
-  abandonedRelationshipChains: number;
-  complexityLoad: number;
-  continuityScore: number;
-  lateWeekDropoff: number;
-};
+export type { LongitudinalPlanningSnapshot } from './longitudinalPlanningTypes';
 
 export type PlannerHistoryProfile = {
   windowSize: number;
