+++ client/src/components/meal-planner/planner-adaptation/longitudinalPlanningTypes.ts
@@ -1,0 +1,18 @@
+export type LongitudinalPlanningSnapshot = {
+  id: string;
+  createdAt: string;
+  weekKey: string;
+  completedMeals: number;
+  skippedMeals: number;
+  repeatedMeals: number;
+  prepCompletionRate: number;
+  groceryCompletionRate: number;
+  hydrationAdherence: number;
+  readinessAverage: number;
+  fatigueAverage: number;
+  successfulRelationshipChains: number;
+  abandonedRelationshipChains: number;
+  complexityLoad: number;
+  continuityScore: number;
+  lateWeekDropoff: number;
+};
