+++ docs/planner-circular-dependency-report.md
@@ -1,0 +1,33 @@
+# Meal Planner Circular Dependency Hardening Report
+
+Generated during the circular dependency hardening pass for `client/src/components/meal-planner`.
+
+## Verification Command
+
+```bash
+npx madge --extensions ts,tsx --circular client/src/components/meal-planner
+```
+
+## Initial Cycles Found
+
+1. `planner-adaptation/adaptationTypes.ts` → `personality-modeling/personalityTypes.ts` → `planner-adaptation/adaptationTypes.ts`
+2. `meal-relationships/relationshipGraph.ts` → `meal-relationships/ingredientRelationships.ts` → `plannerGroceryUtils.ts` → `meal-relationships/relationshipGraph.ts`
+3. `auto-planner/autoPlannerOptimizationEngine.ts` → `planner-objectives/plannerObjectiveEngine.ts` → `planner-objectives/plannerObjectiveScoring.ts` → `auto-planner/autoPlannerRhythmEngine.ts` → `auto-planner/autoPlannerLifestyleAnalysis.ts` → `auto-planner/autoPlannerOptimizationEngine.ts`
+
+## Cycles Resolved
+
+1. Extracted `LongitudinalPlanningSnapshot` into `planner-adaptation/longitudinalPlanningTypes.ts` so personality model types no longer import back through adaptation aggregate types.
+2. Pointed ingredient relationship analysis directly at the foundational ingredient normalization utility instead of importing through grocery aggregation utilities.
+3. Pointed lifestyle analysis directly at foundational auto-planner metrics instead of importing through the optimization engine facade.
+
+## Remaining Cycles
+
+None detected by Madge in `client/src/components/meal-planner` after this pass.
+
+## Final Result
+
+Madge result after the refactor:
+
+```text
+✔ No circular dependency found!
+```
